Avoid ambiguity in print type. There is no portable support yet in Aften for using PRId64 and the like.
git-svn-id: https://aften.svn.sourceforge.net/svnroot/aften@278 ef0d8562-5c19-0410-972e-841db63a069c

+++ util/wavrms.c
@@ -140,7 +140,7 @@ main(int argc, char **argv)
         nr = wavfile_read_samples(&wf, buf, frame_size);
     }
     avg_rms /= avg_cnt;
-    printf("Dialnorm: -%lu dB\n", MIN(avg_rms, 31));
+    printf("Dialnorm: -%d dB\n", (int)MIN(avg_rms, 31));
 
     free(buf);
     fclose(fp);
